fix: clean up latexdiff wrapper validation
